modificate alcune import e rimossi commenti superflui

+++ src/main/java/it/unina/ingSw/cineMates20/utils/Resources.java
@@ -1,5 +1,7 @@
 package it.unina.ingSw.cineMates20.utils;
 
+import it.unina.ingSw.cineMates20.view.MessageDialog;
+
 import java.io.*;
 import java.nio.file.Files;
 import java.nio.file.Paths;
@@ -17,7 +19,7 @@ public class Resources {
             properties.loadFromXML(Resources.class.getClassLoader().getResourceAsStream(NAME_RESOURCES_FIlE));
             secretKeysProperties.load(new FileInputStream("secretKeys.properties"));
         } catch (Exception e) {
-            e.getStackTrace();
+            MessageDialog.error("Si è verificato un errore", "Si è verificato un errore, riprova riavviando l'applicativo.");
         }
     }
 
@@ -27,12 +29,13 @@ public class Resources {
         return properties.getProperty(name.toString());
     }
 
-    public static void removeHashEmail() {
+    public static void removeEmail() {
         File file = new File("src/main/resources/loggedUser.txt");
-        file.delete();
+        if(file.exists())
+            file.delete();
     }
 
-    public static void storeHashEmail(String emailHash) {
+    public static void storeEmail(String email) {
         File file = new File("src/main/resources/loggedUser.txt");
 
         try {
@@ -40,21 +43,17 @@ public class Resources {
                 if(!file.createNewFile())
                     return;
             FileWriter writer = new FileWriter(file);
-            writer.write(emailHash);
+            writer.write(email);
             writer.close();
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
+        } catch (IOException ignore) {}
     }
 
-    public static String getEmailHash() {
+    public static String getEmail() {
         try {
             File file = new File("src/main/resources/loggedUser.txt");
             if(file.exists())
                 return Files.readString(Paths.get("src/main/resources/loggedUser.txt"));
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
+        } catch (IOException ignore) {}
         return null;
     }
